Pdf document describing formal aspects of a fractal line pied de poule
To formalise the fractal line pied-de-poule as described by Feijs and Toeters in their Bridges 2015 paper
(archive.bridgesmathart.org/2015/bridges2015-223.pdf) we deploy Lindenmayer rules using two main formal symbols.
These are F and F where the former is interpreted as a turtle graphics forward step and where the latter can
either expand to a full pied-de-poule-like zigzag line or just act as a forward step. The idea that the turtle
writes either thin lines or thick lines (implemented by recursion) is reflected in the typography of these two
symbols. For the aesthetic and cultural implications of the formalised constructs, we refer to the forthcoming
paper by Loe Feijs, to appear in the Journal of Humanistic Mathematics in 2020. Here we explore the formal properties and related proofs.